+++ package.json
@@ -26,8 +26,10 @@
     "react-hook-form": "^7.34.0",
     "react-loading": "^2.0.3",
     "react-modal": "^3.15.1",
+    "react-router-dom": "^6.4.0",
     "sharp": "^0.30.7",
     "simple-react-calculator": "^0.4.1",
+    "sweetalert": "^2.1.2",
     "swr": "^1.3.0"
   },
   "devDependencies": {
